Do not unnecessarily cast format parameters
Types like strong enumerations and std::uint8_t are gracefully handled
by the formatter.

+++ fly/config/config_manager.cpp
@@ -30,7 +30,7 @@ ConfigManager::ConfigManager(
             break;
 
         default:
-            LOGE("Unrecognized configuration type: {}", static_cast<int>(file_type));
+            LOGE("Unrecognized configuration type: {}", file_type);
             break;
     }
 }
